refactor: MoE + MoEDispatch own perf data (AIC-537 / Phase 4.1)
Phase 4 — moves ``MoE`` and ``MoEDispatch`` out of
``operations/_legacy.py`` into a new ``operations/moe.py``.

``MoE`` owns four CSV tables, all bound on its class:
- ``_moe_data`` — regular MoE
- ``_moe_low_latency_data`` — TRT-LLM low-latency NVFP4 kernel; loaded
  from the same CSV as the regular table since ``load_moe_data`` is the
  only loader that returns a tuple ``(default, low_latency)``.
- ``_wideep_context_moe_data`` — SGLang WideEP context MoE (SGLang-only)
- ``_wideep_generation_moe_data`` — SGLang WideEP generation MoE
  (SGLang-only)

``MoEDispatch`` owns two SGLang DeepEP tables:
- ``_wideep_deepep_normal_data``
- ``_wideep_deepep_ll_data``

On non-SGLang backends, the WideEP / DeepEP caches bind to ``None`` and
consumers must guard before dereference. ``PerfDatabase.query_moe`` /
``query_wideep_deepep_ll`` / ``query_wideep_deepep_normal`` reduce to
one-line delegations to the new classmethods.

The legacy ``MoEDispatch.query`` body is preserved verbatim — it
dispatches across NCCL, CustomAllReduce, TRT-LLM AllToAll, and SGLang
DeepEP paths based on backend + ``_sm_version`` + ``_moe_backend``.
Most of those calls go through ``database.query_*`` helpers that other
migrated ops already own.

Eager ``MoE.load_data(self)`` / ``MoEDispatch.load_data(self)`` calls
added in ``PerfDatabase.__init__`` alongside the existing GEMM /
ContextAttention / GenerationAttention transition shims.
ISSUE-16 cleanup retires all of them.
Verification:
- 670 unit tests pass + 1 skipped (same as baseline)
- 38 regression snapshots bit-identical + 6 skipped (same as baseline)
- ruff check + format clean
