+++ src/aiconfigurator/sdk/operations/_legacy.py
@@ -278,427 +278,6 @@ class TrtLLMWideEPMoEDispatch(Operation):
     def get_weights(self, **kwargs):
         """MoE dispatch has no weight memory."""
         return 0.0
-
-
-class MoE(Operation):
-    """
-    MoE operation with power tracking.
-    """
-
-    def __init__(
-        self,
-        name: str,
-        scale_factor: float,
-        hidden_size: int,
-        inter_size: int,
-        topk: int,
-        num_experts: int,
-        moe_tp_size: int,
-        moe_ep_size: int,
-        quant_mode: common.MoEQuantMode,
-        workload_distribution: str,
-        attention_dp_size: int,
-        is_context: bool = True,
-        is_gated: bool = True,
-        **kwargs,
-    ) -> None:
-        super().__init__(name, scale_factor)
-        self._hidden_size = hidden_size
-        self._inter_size = inter_size
-        self._quant_mode = quant_mode
-        self._topk = topk
-        self._num_experts = num_experts
-        self._moe_tp_size = moe_tp_size
-        self._moe_ep_size = moe_ep_size
-        self._attention_dp_size = attention_dp_size
-        self._workload_distribution = workload_distribution
-        self._is_context = is_context
-        self._is_gated = is_gated
-        self._moe_backend = kwargs.get("moe_backend")
-        self._enable_eplb = kwargs.get("enable_eplb", False)
-        # 3 GEMMs for gated (gate, up, down), 2 GEMMs for non-gated (up, down)
-        num_gemms = 3 if is_gated else 2
-        self._weights = (
-            self._hidden_size
-            * self._inter_size
-            * self._num_experts
-            * quant_mode.value.memory
-            * num_gemms
-            // self._moe_ep_size
-            // self._moe_tp_size
-        )
-
-    def query(self, database: PerfDatabase, **kwargs) -> PerformanceResult:
-        """Query MoE latency with energy data."""
-        # attention dp size will scale up the total input tokens.
-        x = kwargs.get("x") * self._attention_dp_size
-        overwrite_quant_mode = kwargs.get("quant_mode")
-        quant_mode = self._quant_mode if overwrite_quant_mode is None else overwrite_quant_mode
-
-        result = database.query_moe(
-            num_tokens=x,
-            hidden_size=self._hidden_size,
-            inter_size=self._inter_size,
-            topk=self._topk,
-            num_experts=self._num_experts,
-            moe_tp_size=self._moe_tp_size,
-            moe_ep_size=self._moe_ep_size,
-            quant_mode=quant_mode,
-            workload_distribution=self._workload_distribution,
-            is_context=self._is_context,
-            moe_backend=self._moe_backend,
-            is_gated=self._is_gated,
-            enable_eplb=self._enable_eplb,
-        )
-
-        return PerformanceResult(
-            float(result) * self._scale_factor,
-            energy=result.energy * self._scale_factor,
-            source=getattr(result, "source", "silicon"),
-        )
-
-    def get_weights(self, **kwargs):
-        return self._weights * self._scale_factor
-
-
-# a comm op to deduce the communication cost of MoE
-class MoEDispatch(Operation):
-    """
-    MoE dispatch operation. For fine grained moe dispatch
-    """
-
-    def __init__(
-        self,
-        name: str,
-        scale_factor: float,
-        hidden_size: int,
-        topk: int,
-        num_experts: int,
-        moe_tp_size: int,
-        moe_ep_size: int,
-        attention_dp_size: int,
-        pre_dispatch: bool,
-        enable_fp4_all2all: bool = True,
-        **kwargs,
-    ) -> None:
-        super().__init__(name, scale_factor)
-        self._hidden_size = hidden_size
-        self._topk = topk
-        self._num_experts = num_experts
-        self._moe_tp_size = moe_tp_size
-        self._moe_ep_size = moe_ep_size
-        self._attention_dp_size = attention_dp_size
-        self._weights = 0.0
-        self._enable_fp4_all2all = enable_fp4_all2all
-        self._pre_dispatch = pre_dispatch
-        self.num_gpus = self._moe_ep_size * self._moe_tp_size
-        self._attention_tp_size = moe_tp_size * moe_ep_size // self._attention_dp_size
-        self._sms = kwargs.get("sms", 12)
-        self._moe_backend = kwargs.get("moe_backend")
-        self._is_context = kwargs.get("is_context", True)
-        self._scale_num_tokens = kwargs.get("scale_num_tokens", 1)
-        self._quant_mode = kwargs.get("quant_mode")
-        self._reduce_results = kwargs.get("reduce_results", True)
-
-    def query(self, database: PerfDatabase, **kwargs) -> PerformanceResult:
-        num_tokens = kwargs.get("x")
-        volume = num_tokens * self._hidden_size
-        _sm_version = database.system_spec["gpu"].get("sm_version", -1)
-        _num_gpus_per_node = database.system_spec["node"]["num_gpus_per_node"]
-        _node_num = self.num_gpus / _num_gpus_per_node
-
-        if self._quant_mode is not None:
-            _quant_compress = self._quant_mode.value.memory / 2.0
-        else:
-            _quant_compress = 0.25
-
-        if database.backend == common.BackendName.trtllm.value:
-            assert self._attention_tp_size == 1 or self._attention_dp_size == 1, (
-                "trtllm does not support TP>1 and DP>1 for attn simultaneously"
-            )
-            if _sm_version == 100:
-                logger.debug("MoEDispatch: In trtllm SM100 execution path")
-
-                _alltoall_backends = {"CUTLASS", "TRTLLM"}
-                backend_supports_alltoall = self._moe_backend is None or self._moe_backend.upper() in _alltoall_backends
-                is_nvl72 = _num_gpus_per_node >= 72
-                enable_alltoall = (
-                    backend_supports_alltoall and self._attention_dp_size > 1 and self._moe_tp_size == 1 and is_nvl72
-                )
-
-                # Quantize-aware communication volume.
-                # When quant_mode is known, compute compressed volume:
-                #   nvfp4: volume/4 + scale_factor volume
-                #   fp8:   volume/2
-                #   others / unknown: full volume (BF16)
-                quant_mode = self._quant_mode
-                if quant_mode is not None and quant_mode == common.MoEQuantMode.nvfp4:
-                    dispatch_x_volume = volume / 4
-                    dispatch_sf_volume = volume / 4 / 8
-                elif quant_mode is not None and quant_mode in (common.MoEQuantMode.fp8, common.MoEQuantMode.fp8_block):
-                    dispatch_x_volume = volume / 2
-                    dispatch_sf_volume = 0
-                else:
-                    dispatch_x_volume = volume
-                    dispatch_sf_volume = 0
-
-                if enable_alltoall and quant_mode is None:
-                    raise ValueError("MoEDispatch requires quant_mode when TRTLLM alltoall path is enabled.")
-
-                if self._pre_dispatch:
-                    if enable_alltoall:
-                        dispatch_result = database.query_trtllm_alltoall(
-                            op_name="alltoall_dispatch",
-                            num_tokens=num_tokens,
-                            hidden_size=self._hidden_size,
-                            topk=self._topk,
-                            num_experts=self._num_experts,
-                            moe_ep_size=self._moe_ep_size,
-                            quant_mode=quant_mode,
-                            moe_backend=self._moe_backend,
-                        )
-                        comm_latency = float(dispatch_result)
-                    elif self._attention_dp_size > 1:
-                        all_gather_volume = (dispatch_x_volume + dispatch_sf_volume) * self._attention_dp_size
-                        comm_latency = database.query_nccl(
-                            common.CommQuantMode.half, self.num_gpus, "all_gather", all_gather_volume
-                        )
-                    elif self._attention_tp_size > 1:
-                        if self._reduce_results:
-                            if _num_gpus_per_node == 72 and self.num_gpus > 4:
-                                comm_latency = database.query_nccl(
-                                    common.CommQuantMode.half, self.num_gpus, "all_reduce", volume
-                                )
-                            else:
-                                comm_latency = database.query_custom_allreduce(
-                                    common.CommQuantMode.half, self.num_gpus, volume
-                                )
-                        else:
-                            comm_latency = 0
-                    else:
-                        comm_latency = 0
-                else:
-                    if enable_alltoall:
-                        combine_result = database.query_trtllm_alltoall(
-                            op_name="alltoall_combine",
-                            num_tokens=num_tokens,
-                            hidden_size=self._hidden_size,
-                            topk=self._topk,
-                            num_experts=self._num_experts,
-                            moe_ep_size=self._moe_ep_size,
-                            quant_mode=quant_mode,
-                            moe_backend=self._moe_backend,
-                        )
-                        comm_latency = float(combine_result)
-                    elif self._attention_dp_size > 1:
-                        comm_latency = database.query_nccl(
-                            common.CommQuantMode.half,
-                            self.num_gpus,
-                            "reduce_scatter",
-                            volume * self._attention_dp_size,
-                        )
-                    elif self._attention_tp_size > 1:
-                        if self._reduce_results:
-                            if _num_gpus_per_node == 72 and self.num_gpus > 4:
-                                comm_latency = database.query_nccl(
-                                    common.CommQuantMode.half, self.num_gpus, "all_reduce", volume
-                                )
-                            else:
-                                comm_latency = database.query_custom_allreduce(
-                                    common.CommQuantMode.half, self.num_gpus, volume
-                                )
-                        else:
-                            comm_latency = 0
-                    else:
-                        comm_latency = 0
-            else:  # sm < 100 or > 100 (for now)
-                logger.debug("MoEDispatch: In trtllm SM<100 or >100 execution path")
-                if self._pre_dispatch:
-                    if self._attention_tp_size > 1:  # tp>1, use allreduce
-                        # to do: custom allreduce
-                        comm_latency = database.query_custom_allreduce(common.CommQuantMode.half, self.num_gpus, volume)
-                    elif self._attention_dp_size > 1:
-                        comm_latency = database.query_nccl(
-                            common.CommQuantMode.half,
-                            self.num_gpus,
-                            "all_gather",
-                            volume * self._attention_dp_size,
-                        )
-                    else:
-                        comm_latency = 0
-                else:
-                    if self._attention_tp_size > 1:  # tp>1, use allreduce
-                        # to do: custom allreduce
-                        comm_latency = database.query_custom_allreduce(common.CommQuantMode.half, self.num_gpus, volume)
-                    elif self._attention_dp_size > 1:
-                        comm_latency = database.query_nccl(
-                            common.CommQuantMode.half,
-                            self.num_gpus,
-                            "reduce_scatter",
-                            volume * self._attention_dp_size,
-                        )
-                    else:
-                        comm_latency = 0
-        elif database.backend == common.BackendName.vllm.value:
-            assert self._moe_tp_size == 1 or self._moe_ep_size == 1, (
-                "vllm does not support MoE TP and MoE EP at the same time"
-            )
-
-            comm_latency = 0
-
-            # Add allreduce latency when TP > 1
-            if self._attention_tp_size > 1:
-                comm_latency += database.query_custom_allreduce(common.CommQuantMode.half, self.num_gpus, volume)
-
-            if self._attention_dp_size > 1:
-                comm_latency += database.query_nccl(
-                    common.CommQuantMode.half,
-                    self.num_gpus,
-                    "all_gather" if self._pre_dispatch else "reduce_scatter",
-                    volume * self._attention_dp_size,
-                )
-        elif database.backend == common.BackendName.sglang.value:
-            if self._moe_backend == "deepep_moe":
-                logger.debug("MoEDispatch: In SGLang DeepEP execution path")
-                num_tokens = num_tokens // self._scale_num_tokens
-                if self._is_context:
-                    comm_latency = database.query_wideep_deepep_normal(
-                        node_num=_node_num,
-                        num_tokens=num_tokens,
-                        num_experts=self._num_experts,
-                        topk=self._topk,
-                        hidden_size=self._hidden_size,
-                        sms=self._sms,
-                    )
-                else:
-                    comm_latency = database.query_wideep_deepep_ll(
-                        node_num=_node_num,
-                        num_tokens=num_tokens,
-                        num_experts=self._num_experts,
-                        topk=self._topk,
-                        hidden_size=self._hidden_size,
-                    )
-            else:
-                logger.debug("MoEDispatch: In SGLang non-DeepEP execution path")
-                combined_attention_tpdp = self._attention_tp_size > 1 and self._attention_dp_size > 1
-                if self._pre_dispatch:
-                    if combined_attention_tpdp:
-                        # Matches SGLang DP attention: shard across attention TP, then gather across the full TP world.
-                        comm_latency = database.query_nccl(
-                            common.CommQuantMode.half,
-                            self._attention_tp_size,
-                            "reduce_scatter",
-                            volume,
-                        )
-                        comm_latency += database.query_nccl(
-                            common.CommQuantMode.half,
-                            self.num_gpus,
-                            "all_gather",
-                            volume * self._attention_dp_size,
-                        )
-                    elif self._attention_tp_size > 1:  # tp>1, use allreduce
-                        # to do: custom allreduce
-                        comm_latency = database.query_custom_allreduce(common.CommQuantMode.half, self.num_gpus, volume)
-                    elif self._attention_dp_size > 1:
-                        comm_latency = database.query_nccl(
-                            common.CommQuantMode.half,
-                            self.num_gpus,
-                            "all_gather",
-                            volume * self._attention_dp_size,
-                        )
-                    else:
-                        comm_latency = 0
-                else:
-                    if combined_attention_tpdp:
-                        # Reverse path: reduce-scatter across the full TP world, then rebuild each attention TP group.
-                        comm_latency = database.query_nccl(
-                            common.CommQuantMode.half,
-                            self.num_gpus,
-                            "reduce_scatter",
-                            volume * self._attention_dp_size,
-                        )
-                        comm_latency += database.query_nccl(
-                            common.CommQuantMode.half,
-                            self._attention_tp_size,
-                            "all_gather",
-                            volume,
-                        )
-                    elif self._attention_tp_size > 1:  # tp>1, use allreduce
-                        # to do: custom allreduce
-                        comm_latency = database.query_custom_allreduce(common.CommQuantMode.half, self.num_gpus, volume)
-                    elif self._attention_dp_size > 1:
-                        comm_latency = database.query_nccl(
-                            common.CommQuantMode.half,
-                            self.num_gpus,
-                            "reduce_scatter",
-                            volume * self._attention_dp_size,
-                        )
-                    else:
-                        comm_latency = 0
-        else:  # other backends
-            raise NotImplementedError(f"MoEDispatch: Not implemented for backend {database.backend}")
-
-        scaled = comm_latency * self._scale_factor
-        return PerformanceResult(
-            float(scaled),
-            energy=getattr(scaled, "energy", 0.0),
-            source=getattr(scaled, "source", "empirical"),
-        )
-
-    def get_weights(self, **kwargs):
-        return self._weights * self._scale_factor
-
-    def query_ideal(self, database: PerfDatabase, **kwargs):
-        """
-        Ideal communication cost for MoE dispatch. For reference only.
-        """
-        num_tokens = kwargs.get("x")
-        volume = num_tokens * self._hidden_size
-
-        if self._pre_dispatch:
-            reduce_scatter1_v = volume / self.num_gpus
-            reduce_scatter1_num_gpus = self._attention_tp_size
-
-            all2all1_v = volume * self._topk / self.num_gpus
-            all2all1_num_gpus = self.num_gpus
-
-            allgather1_v = volume / self._moe_tp_size
-            allgather1_num_gpus = self._moe_tp_size
-
-            comm_latency = (
-                database.query_nccl(
-                    common.CommQuantMode.half,
-                    reduce_scatter1_num_gpus,
-                    "reduce_scatter",
-                    reduce_scatter1_v,
-                )
-                + database.query_nccl(common.CommQuantMode.half, all2all1_num_gpus, "alltoall", all2all1_v)
-                + database.query_nccl(common.CommQuantMode.half, allgather1_num_gpus, "all_gather", allgather1_v)
-            )
-        else:
-            reduce_scatter2_v = volume
-            reduce_scatter2_num_gpus = self._moe_tp_size
-
-            all2all2_v = volume * self._topk / self.num_gpus
-            all2all2_num_gpus = self.num_gpus
-
-            allgather2_v = volume / self.num_gpus
-            allgather2_num_gpus = self._attention_tp_size
-
-            comm_latency = (
-                database.query_nccl(
-                    common.CommQuantMode.half,
-                    reduce_scatter2_num_gpus,
-                    "reduce_scatter",
-                    reduce_scatter2_v,
-                )
-                + database.query_nccl(common.CommQuantMode.half, all2all2_num_gpus, "alltoall", all2all2_v)
-                + database.query_nccl(common.CommQuantMode.half, allgather2_num_gpus, "all_gather", allgather2_v)
-            )
-
-        return comm_latency * self._scale_factor
-
-
 class FallbackOp(Operation):
     """
     Try a primary operation first; if it raises PerfDataNotAvailableError,
